Fix nosuchmethod errors when running on different java versions
See https://www.baeldung.com/maven-java-version#java9

--

By default, the -source and -target options don't guarantee a cross-compilation.
This means that we cannot run our application on older versions of the platform. Additionally, to compile and run the programs for older Java versions, we also need to specify -bootclasspath option.
To cross-compile correctly, the new -release option replaces three flags: -source, -target and -bootclasspath.
(...)
And for the maven-compiler-plugin starting from the 3.6 version we can write:
<plugin>
    <groupId>org.apache.maven.plugins</groupId>
    <artifactId>maven-compiler-plugin</artifactId>
    <version>3.8.0</version>
    <configuration>
        <release>7</release>
    </configuration>
</plugin>