Fix caching bugs by consistently defining hashCode() and equals().
All BLSpanQuery classes should have these, as they are used to test if a
search is in the cache already. If one class is missing these methods or
they are incorrectly implemented, the super class version is used, which
might not look at all relevant instance variables. This was the case for
SpanQueryFiltered for example, causing the same pattern with different
metadata filter to return the same (cached).