Fix groups not accurately reporting maxstats
Normally we can't just copy maxstats from source hits to derived hits, as the source hits may be vacated from the cache, or may still be processing at the time the derived hits (window/sample/etc.) are created. But for groups this is different, as the source hits need to be completely read/processed before constructing groups (or inaccurate results are shown).
So in this case, we can copy the HitsStats and DocsStats from the hits.

The upside of this is that Blacklab-Server now accurately reports whether the query was limited for grouped results.