Greatly improve startup time and passive memory usage
Eliminated the need for TermsReader to calculate the collation keys for
all terms on startup. This improves startup time by ~5-6x and halves
memory usage.
This comes at a small price for finding a string's term id, which now
needs a binary search instead of a hash and direct lookup. However this
is almost never used at runtime, and certainly not in any hot path, so
the performance difference should not be noticable.