Fix trailing tokens being counted when grouping all tokens
In the forward index, BlackLab seems to store one more token than the actual length of the document.
The token frequency grouping path did not account for this, causing one extra token to be counted for every document in a corpus.
The tokens in question have null values for every annotation (except, perhaps, punctuation).
We now skip these tokens.