Revert "replace collators with a faster more memory efficient implementation"
This reverts commit 91c9fce57c37437c0502ed95533faba75e3cb730.

The ICU collator implementation has some subtle ordering differences compared to the default java implementation.
This causes issues when using indexes from before the change. The first issue found is that retrieving group contents sometimes fails. The cause is that some implementations of PropertyValue use the sortPosition of an annotation's term, and with a different sorting of terms than stored on disk (for an old index) this index no longer matches and issues arise.