Add option "keep" option "all" to split action, keeps original + all splits
This can be used to (for example), split multiple lemmata for a single token, but also keep the joined lemma.
The original (unsplit) value is indexed at the first position, so that is what appears when retrieving from the forward index.

Also use precompiled pos parse regex for slightly faster indexing when using action parsePos.