Fix not all hits always counted when grouping/sorting
The issue was that in large corpora (those containing multiple leaves),
HitsFromQueryParallel would always launch multiple threads, even when
only 1 extra hit needs to be read.
All of these threads would always count/process at least 1 hit before
aborting, however only one of the threads could actually "store" the
hit. This caused hits to be "skipped" and thus the count and process
counts to begin to differ. This was reported in the corpus-frontend as a
query being "limited" (because not all hits could be stored, which is
usually only the case when the limit for storing hits has been lowered
in the blacklab-server configuration file).

The solution is to always pre-fetch one hit, simply do nothing if it
cannot be counted. The next time(s) a hit is requested, the prefetched
hit has another chance to be processed, until it eventually gets used.

Tangentially related to these two issues
(because of the way this present, it may have looked like a more serious bug):
Issue #177
Issue #181