Fix: buffer overflow in TermsWriter with >134M unique terms.
Solved by correclty using long calculation when determining the size
on disk for the sort buffers.

Please note that there cannot be more than 2G unique terms with the
current way the code is set up. If this is a problem because you're
indexing unique word IDs, try disabling the forward index.