Don't abort running searches because no threads are available.
ExecutorService will queue the search until a thread is available.
Aborting based on number of searches doesn't work, partially because we
also have SpansReaders from HitsFromQueryParallel that aren't counted.
And this doesn't solve the deadlock issues when e.g. a hits task is
queued and a sort task is running that needs its results.