Do remove abandoned counts when aborting them. Track cancelled search.
Counts are aborted more aggressively than searches (say after 30s), so a
user might press the back button and want to see the count again. Don't
penalize them for our quick-triggered culling of counts.

When a search is aborted, the future is set to null so the memory can be
freed, but this means we should keep track of the fact that it was
cancelled separately. We do this now.