HitGroupsTokenFrequencies: use lengthfield, not FI, for doc length.
Different annotations may have a different number of tokens indexed. For
example, there can be punctuation or closing tags after the last word,
which is why those annotations always have an "extra closing token".

Not all annotations may have this though (e.g. it doesn't make sense for
word, lemma, etc.). Note that the document length stored in the length
field will always include this extra closing token, so if we want the
real document length, we subtract one.

Because the forward index may have different numbers of tokens for
different annotations, we look at the length field, not at the number of
tokens in the FI. The number of tokens in the FI is always at least the
number of words in the document, and at most one more.