Start work on custom Codec (still unused).
The goal of this custom codec is to incorporate all
external files (forward index, content store, indexmetadata,
version files) in the Lucene index.

This is needed for using BlackLab with (distributed) Solr
in the future.

A description of the codec files can be found in:
core/doc/blacklab codec file format.txt