Fixed intermittent crashes using synchronization. (#259)
There would occassionaly be strange crashes while searching, mostly related to reading
from the Lucene index or the forward index.

An example of a query that frequently crashed, especially shortly after starting the server:

/test/hits?patt=%5B%5D&sort=wordleft%3Aword%3Ai%2Cwordright%3Aword%3Ai%2Cfield%3Apid&wordsaroundhit=1
(usecache=no, waitfortotal=true were also passed)

The problems appear to have been introduced when some operations were parallelized, but because
classes weren't clearly marked as threadsafe or not, not all necessary synchronization was done.

This adds synchronization where needed and marks several important classes as
thread-safe (or not). This appears to solve the intermittent crashes.

It would be good to investigate and document the thread-safety of more classes.