Fix off by one in DocResults. Document extra closing token. (#261)
Document length in Lucene and forward index is always reported as one higher due to punctuation being a trailing value. We call this the "extra closing token".

The code wasn't always explicit about what document length meant; sometimes it included the extra closing token, sometimes not. I've tried to document it so it's clear which value we're dealing with.

Also discovered an off-by-one bug in DocResults because we forgot to subtract the extra closing token.