Fix and simplify peeking at count.
Before we used a delegate, but this created problems when other
parts of the code assumed all hits had been gathered, so therefore
the count obtained via another search would be the final count right
away, but because executeInternal handn't been called yet on the
SearchCountFromResults operation, the delegate returned 0 instead
of the final count.

The other parts of the code that make assumptions about related searches
should also be improved to be more explicit about how searches relate
to one another, and what the state of each search is (running, completed, etc.).