Peek without waiting. Immutable Search classes. (#276)
SearchResultsFromCount is now immutable again, as all Search classes should be.

Instead, to be able to follow the running hits/docs count, we now pass a Peekable (implemented by SearchCacheEntry) to executeInternal(). This method can use the object returned by Peekable.peek() to give updates on its progress while the search is executing. If null is passed to executeInternal(), you won't be able to follow the running count, but everything should work correctly otherwise.

For counting, because we want to be able return a ResultsStats object right away, even if the "real" count isn't available yet, we use ResultsStatsDelegate. This will return 0 if the real count isn't available and delegate to the real count object when it is. CAUTION: if you call processedTotal() or countedTotal() on the ResultsStatsDelegate, it will need to block until all hits have been processed. All other operations will return right away, however.

This has gotten fairly complex and it would be nice to simplify this in the future, while preserving functionality.