Be explicit about HitsInternalRead and immutable Hits objects.
Where possible, we restrict ourselves to using HitsInternalRead so we know the code
cannot modify this HitsInternal object. We've also renamed HitsList to HitsImmutable
and ensured that this class only has access to the HitsInternalRead interface for this
reason. Classes that do need to mutate their HitsInternal object are now derived from
HitsInternalMutable.