Significant code cleanup.
A whole bunch of code cleanups, including:
- optimize imports (according to [Eclipse import order](https://stackoverflow.com/a/17194980), which is what BlackLab has always used)
- remove unused or commented out code
- simplify code where possible
- make fields final where possible
- convert field to local variable where possible
- remove copyright notices at the top of source files (we don't need to repeat the information from LICENSE.txt everywhere)
- updated dependencies and plugins
- made `mvn site` work a bit better again (and documented how to do it in README.md)
- added formatter settings for IntelliJ (and documented that these exist in README.md)

This generates a one-time large diff, but shouldn't change any semantics (unless I made a mistake in one of my manual edits, of course).

We should try to introduce as few of these issues as possible and do smaller cleanups regularly to avoid these large cleanups in the future.