Fix metadata value frequency reading.
With the alternate way of writing YAML files introduced a while back, numbers would
sometimes be written to indexmetadata.yaml as strings, particularly metdata field value
frequencies. We used intValue() when reading these back, resulting in all of them getting
the default value of 0. We now use asInt() so they are automatically converted from string
to number. Shouldn't be necessary, but it is.

Because of this fix, expected-index-metadata.json had to be updated, as it now correctly
reports the frequencies of the metadata field values.

Also removed unused JsonUtil and tweaked the YamlObjectMapper test.