Don't re-use contextIndices in HitProperty.
contextIndices are an internal helper value and are not immutable, so sharing
them between HitProperty instances can lead to hard to debug issues.
It doesn't make sense anyway, as this doesn't save any memory to speak of.

In this case, usecache=no triggered an issue where contextIndices would seemingly
change during a sort operation. This change fixed that.