Debug mode: originating ip. Subnets.
We now use the originating ip (X-Forwarded-For header if present) to
determine debug mode (otherwise BLS behind a proxy always looks like a local
request).

It is now also possible to use wildcards * and ? in the debug.addresses section
of blacklab-server.yaml to mark entire subnets as debug addresses, e.g.
"172.16.4.*".