Refactor BlackLabIndex, add BlackLabIndexIntegrated.
This cleans up the BlackLabIndex interface a bit, introduces
an abstract base class and a new implementation for the integrated
index.

Some static variables and methods were moved from BlackLabIndexAbstract
to BlackLab and BuiltinAnalyzers, where they fit better.

Trace settings are now dynamic, so they could in theory be changed at runtime.
Before the value was copied from the global config once.

BlackLab.defaultCollator() uses the configured collator now instead of
a hardcoded default.

Still to do is to try to eliminate methods related to forward index id
(fiid) from the BlackLabIndex interface, because this is a implementation-
specific thing that will disappear eventually when all indexes are integrated.

Also makes sure that all BlackLabEngine instances are closed on
shutdown, even if the application forgets to do this.