Refactor ForwardIndex APIs to work with Lucene doc id, not fiid.
The forward index id (fiid) was an implementation-specific number that is no
longer needed with the integrated index, so it makes sense to isolate it in the
relevant implementation classes. It also makes for simpler code and a cleaner
API.

AnnotationForwardIndexExternalAbstract now owns a FiidLookup instance that it will
use to convert Lucene docIds to fiids. FiidLookup is not threadsafe, so synchronization
is used. This should not slow things down as looking up fiids is only a small part of
accessing the forward index (I/O and processing tokens generally takes much longer).