Re-use BLTerms. Access terms terms per-segment.
Before we would instantiate BLTerms multiple times, but this is unnecessary
and causes problems. We now keep one instance per segment and reuse that.

Before ForwardIndexAccessorLeafReaderIntegrated would perform conversion
between segment term ids and global term ids; now it just forwards the
call to BLTerms.

BLTerms right now still does translation to global term ids, but that could
be changed in the future, and this way the implementation is centralized.