Work on indexmetadata.
Document-based approach is probably not a good idea; having an odd document
in the index causes too many issues.

Instead, we should probably write the indexmetadata to the field attributes.
Each index segment gets its own copy. This does mean that the metadata cannot
change after indexing has started, which is a bit of an issue because of
forEach. But we may have to accept that all fields will have to be declared
before use.