Refactor Indexer/BlackLabIndexWriter instantiation. IndexTool option --integrate-external-files (#325)
Instead of letting the Indexer create or open the index, let the client do that and pass the resulting BlackLabIndexWriter to the Indexer. This is more flexible and will make testing easier.

This proved a little tricky because Indexer did some additional things that didn't normally happen when opening an index, such as making sure the documentFormat is set in the metadata. But it should work now.

The entire reason for the above change was the desire for an option in the IndexTool to specify the index format, external or integrated. That should work now as well.

Also fixes "only warn once about relying on default sensitivity".