Revert to storing index metadata in a document.
The segment approach didn't work for an empty index, and doesn't
work well with appending documents later.

The index metadata document shouldn't be seen as an "abnormal" document
to be skipped. It should be seen as a document that is part of our Lucene
index and contains only (special) metadata. We should skip documents that
don't contain the annotated field we're searching; this will automatically
skip the index metadata document as well.

WIP. documentCount still reports one too many documents, but we needed to look
at this anyway (uses maxDoc right now, not live docs?).