Reorganize test/ directory and Docker setup.
- test/saved-responses/ was moved into test/data/.
- input files were moved into test/data/input/
- indices.js expected response files were moved to test/saved-responses/indices/

Previously, test/data was mounted to /input in the BLS testserver container; now it
is mounted to /test-data. This affects the value of the field fromInputFile.

The files to index inside the testserver container are now in /test-data/input/.

Inside the test container, the saved responses were at /app/saved-responses but are
now at /app/data/saved-responses.