Use CustomProps interface for access to all custom properties.
The generic implementation CorpusPropsMap just wraps a Map and is the future.
For legacy support, there will be implementations that delegate to the old deprecated methods like displayName(), description(), etc.

Reasoning: BlackLab does not need access to these and we should not pollute our APIs with properties for all possible applications. We also don't need BlackLab to validate any of these; the applications that use them should do that.