Better integration tests.
Integration tests will now compare entire BLS responses.

Run test/run-ci.sh to run tests, comparing with previously saved responses. A missing response will cause the test to fail.

Run test/run-local.sh to run tests locally. A missing response will cause the test to SUCCEED and the response will be saved.

This makes writing tests quicker and tests more thorough because the
entire response is compared to the previous version. Of course, it is
up to the test writer to check that the saved response is indeed correct
before committing it as "known good".