Look at metadata to decide if we need FI. Fieldattribute.
Only the main sensitivity of some annotations should get a forward index;
now we check that. We don't yet have access to the indexmetadata if appending
to existing index yet, need to figure that out.

We store in the field attribute when the payloads contain is-primary-value
indicators. See PayloadUtils for details.