Unit test with preindexed classic external index.
The repo now includes (in core/src/test/resources/) a classic BlackLab index.
This is used to ensure we don't break file compatibility. We used to only test
by creating an index, then testing that.

Eventually we should test the integrated index format in this way too, but that's still
in development, and file formats are still changing.