Fix failing test
The issue was that we looked at sort positions on disk to quickly determine global term sort order.
However, the sort positions were not deduped for case-insensitive compares:
e.g. the sort position for "The" and "the" were not the same, even though they should logically be, because the capital letter should be ignored.
We now load the sort positions from disk, where they are deduplicated, so now "The" and "the" have the same position