Start abstracting Lucene details from indexing.
When indexing with Solr, we talk to Solr's APIs and structures, so we
need the indexing code to be independent of the concrete code that
builds up a document and adds it to the index.

This implements a first version of that. Currently it only has a Lucene
implementation and there's still several places in the code that assume
Lucene (e.g. how BLFieldType are constructed).