Extract protocol-independent code from BlackLab Server  (#357)
BlackLab Server contained a lot of code that intertwined the protocol-specific stuff (parameter names, URL parsing, file upload handling, DataStream'ing output) with actually executing the requested operation. This is a problem for integrating with Solr.

This PR extracts the protocol-independent code to the WebserviceParams, WebserviceOperations and Result* classes in the new wslib module (the "web service library"). UserRequest was also introduced as a portable way to capture the request and response context, for authentication.

A number of extra integration tests were added to ensure that the output of the various endpoints didn't change.

The structure of the code is still not ideal, especially the large WebserviceParams class and sometimes messy Result* classes. These problems existed already; the existing code was mostly just moved around. It would be good to improve this eventually, but that is not the first priority.

NOTE: an unrelated change is the addition of the AuthMethod interface implemented by the Auth* classes. This means we don't need to use reflection anymore to call the auth object's methods. If you use a custom auth class, it must be updated to implement AuthMethod.