Write missing index file footers. Fixes segment merges.
Segment merge would throw an exception indicating the index was malformed/corrupted. This ended up being  because  we didn't write the require footer at the end of our files.
Lucene doesn't complain of this initially (it probably should...), so we didn't notice. But as soon as a segment merge was triggered (during indexing) an exception would be thrown, aborting the indexing procedure.