Abstract how BlacklabIndex and IndexReader relate (Solr) (#369)
BlackLabIndex used to be in charge of opening IndexReader, but this doesn't work with Solr, which manages its own IndexReaders. This allows us to instantiate a BlackLabIndex from an IndexReader instance when needed (and when none exists yet).

A future challenge is how to finalize BlackLabIndex (Solr doesn't tell us when it opens or closes IndexReaders). In the long term, the goal is for BlackLabIndex to no longer hold resources (such as global term ids) but to only keep that kind of state in the per-segment codec classes.

Use BlackLab.indexFromReader(IndexReader) to wrap the reader in a BlackLabIndex instance (or retrieve the one previously created).