Deal gracefully with interrupted forward index initialization.
This can happen if e.g. a commandline utility completes
before the full initialization is done. The running threads
are interrupted and the forward index remains uninitialized.
If for some reason the program keeps running and tries to use
the forward index, it will simply try to initialize again
(running in the foreground).