Add reverse proxy for Solr. CI test Solr as well.
Because exposing Solr directly to the internet is not recommended from a security standpoint, and to make it 100% API compatible with BLS, a reverse proxy was created that forwards API requests to Solr and returns the response, converted to the requested type (XML/JSON/CSV).

Note that operations that create and modify user corpora have not been implemented yet. These are currently not a high priority for us and will be added at a later stage.

Using this proxy, we now also run the CI tests on the Solr module, to check that it is fully compatible with BLS. The Solr module should still be considered experimental, but it should now be possible to use it in place of BLS.

This PR introduces BLS API versioning. You can specify a default parameters.api in the config file, or pass the api parameter with a request. Valid values are 3.0,4.0,current and experimental. This should allow us to evolve the API gracefully.

Still to do:
- set default API compatiblity (ApiVersion.CURRENT) to 4.0 and update expected test responses (but maybe do this later as it is a breaking change in BLS. Or at least document prominently how to best deal with this)
- ensure proxying to BlackLab Server works as well (only tested Solr so far); although not currently essential functionality for us