Fix two issues with the proxy.
Unexpected failure response should include code,
like all BlackLab error responses do.

When using Jersey to perform a Solr request,
we have to escape { and } in parameter values,
otherwise they will be interpreted as a template
slot to be filled in from a variable, which is
not what we want.