Fix negative numbers not working with 3-byte FI codec. (#408)
We didn't take negative numbers into account when storing 1, 2 or 3-byte integers; we only checked the upper bound. It turns out that we do store -1 in the FI (presumably when there's no value at a position). This worked fine for the 1 or 2 byte versions (Java handles these as normal byte or short data types), but our custom 3-byte integer needed some trickery to make sure -1 wasn't deserialized as 2^24-1. Now fixed, logic extracted to a class, and tested for every possible value.