Proxy: fix docInfo separating out metadata in XML response. (#412)
The BLS API has a quirk where document metadata is intermingled with lengthInTokens and mayView, which are not metadata fields but other document properties.

The proxy uses JAXB and keeps these separate. This wasn't a problem for JSON, which has a custom (de)serializer anyway, but the XML response was incorrect because of this. Now fixed with a custom XmlAdapter.

* Fix docInfo separating out metadata in XML response.

* docInfo pid attribute.