Fix an issue with opening large indices on windows
When the index was copied from non-windows to windows systems, the block size of terms and forward index files would be larger than the file map that was opened, causing buffer underruns.
Decouple the max block size to read and write, because opening a large file is never an issue.