Make QueryTool suitable for simple comparative testing.
In addition to the -f option to specify a command file, you can
now specify --mode correctness to show results but no timings and
a larger default page size, or --mode performance to show timings
but no results. --mode all shows both results and timings (the
default in interactive mode).