Fix SpansAnd/Repetition bugs, others. Testing, docs.
SpansAnd didn't deal correctly with clauses that produce hits with the same doc, start and end (but different match info). Now it has a default, slower version that correctly deals with this, and the previous version (SpansAndSimple) is used where possible.

SpansRepetition had a bug that skipped certain matches if the repetition consisted of non-consecutive hits. Again, there's now a slower implementation that correctly deals with this and SpansRepetitionSimple is the previous version that will be used when we know this problem cannot occur. A new guarantee, hitsCanOverlap, was added for this.

Spans classes were documented in the BlackLab internals doc, and QueryTool was made more suitable for larger-scale correctness testing.

SpansInBucketsPerDocumentSorted now uses sort indexes to sort hits, so match info is also sorted along with the starts/ends. Incurs a slight performance hit, but probably not too bad.

Additional tests were added for the above.