Make sure the appropriate uniqueness is used.
Meaning either unique doc, start and end, or unique
doc, start, end and match info. Mostly we use the last
uniqueness requirement, which means there can still be multiple
hits with the same doc, start and end. In a few cases, we
use the version where only doc, start and end have to be unique,
for exampe to decide what optimization we can use.