Makefile to make tagging Docker images easier (#433)
See the Makefile for details. Use VERSION variable to specify the image tag (defaults to "latest", as before). Use -commit variants for targets (e.g. release-commit) to use commit hash as image tag.

Removed the Compose files for the proxy and solr modules as these configurations are now in the main Compose file.