Implement SpanWeight.isCacheable().
Almost all of our SpanWeight objects should be cacheable as they don't
rely on DocValues or global statistics etc. The only we're not sure about is
SpanWeightFiSeq, which might run into trouble with the global forward index API
which we plan to remove eventually. For now, we won't cache SpanWeightFiSeq.