Index relations at source. Fixes, assertions.
We switched to indexing relations at the source because the relation operator yields the source, so we won't have to sort the hits in that case; they will already be sorted.

A matching bug was fixed. Many assertions were added to the Spans classes to make detecting and tracking down matching bugs easier.