Enable regex complement for relations matching.
Lucene's optional ~ operator can be used to find relations
whose type doesn't match something. Useful for exclusions.

We haven't enabled this everywhere because NFA matching would also
have to support this operator. Right now the regexes are assumed to
be compatible, which is not strictly true (e.g. escaping rules for Lucene
are stricter than for Java's regex engine), but this would present real
challenges.