Fix nullptr when requesting index info
It happens when annotatedfield doesn't contain any annotations.
When indexing linked metadata, a dummy annotatedField "metadata" is created. Several parts in the code check for this and ignore it. But IndexMetadataExternal (and perhaps internal too?) let it through. This seems to be required because the IndexMetadata object is used to pre-create the contentstore. If we don't do that, the content store for the metadata won't exist, and the indexing process will throw an exception.

So for now, just let it through, and ignore it when serializing. A more thorough solution might have to be implemented in the future. (We probably DO want to have the field exist in the IndexMetadata object, or indexing new documents into the same corpus later won't work?)