Fix not closing swap file (Saxon).
The Saxon DocIndexer temporarily writes the file to index to disk
if it's very large to save memory, reading it back when it's time
to write it to the content store. However, the FileWriter wasn't
closed when doing this, which caused part of the document not to be
written to disk.