SpansRelations only decodes payload if needed. Cleanup.
In some cases, SpansRelation can move past some matches
without decoding its payload, saving time. This may be rare though.

Specifically, spanMode has to be SOURCE, and the relation will have
to be rejected based only on its source start position (which is
just the position it was indexed at).

This could be improved if we could guarantee that our clause never
matches root relations (because weeding those out required decoding the
payload). Maybe we could encode the fact that it's a root relation into
the term as well? Might be a bit hacky though.

Squashed commit of the following:

commit 5ca92322e85943766dc4a67d252e4b3b579ac340
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Sep 25 13:40:11 2023 +0200

    SpansRelations only decodes payload if needed.

commit 7bb05e94d9b1a9861f46b04ce08f4a03e4b013f0
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Mon Sep 25 12:00:33 2023 +0200

    Always index relations at source.