Fix edge-case with for-each subannotations
When a subannotation forEach had the same valuePath as the parentAnnotation,
the parent value was re-used, even though the _actual_ valuePath was the forEachPath,
so the evaluation context in xml was not equal at all.

Ex:
<main value="something">
  <sub value="subvalue" name="subname"/>
  <sub>...</sub>
</main>

- name: main
  basePath: main
  valuePath: "@value"
  subannotations:
  - forEachPath: sub
    valuePath: "@value"
    namePath: "@name"

In this case, the subannotation would index value "something" instead of "subvalue".
Because the indexer erroneously assumed that both 'valuePath' properties "@value" would result in the same string.