Fix context=s and auto-capturing of spans.
When you perform a query like 'the' within <s/>, the
span will now be automatically captured as 's'.

context=s automatically adds within <s/> to the query,
so it can return the requested context, but the query
reported in the response is now the user's original
query, not the query with the within clause added.
explain=yes does report the full query that's actually
executed.