Index relation twice; use fastest when searching.
Relations with attributes (mostly inline tags for now) will be
indexed once with their attributes and once without. When searching,
if you're not filtering on attributes, we will explicitly search
for the indexed term without attributes, which is faster.

Also strip namespace prefix from attribute while indexing. On the
one hand, the BCQL parser currently doesn't allow colon in attribute
name. On the other hand, attributes with the same name might clash.