Don't include attributes in match info.
The problem is that we index relations and tags twice now,
once with and once without attributes, to improve search speed.
If we always wanted to include attribute values in the match info,
we'd have to store them in the payload, which would require a lot
of extra space for too little (?) gain.