Make sure linked doc can find meta values.
E.g. when indexing CGN, the FoLiA links to the IMDI doc with the
metadata. That uses fromInputFile to determine the Collection,
but we need access to that value, which is stored in the linking
DocIndexer (the one indexing the FoLiA doc). We now allow that
access so linked documents can use metadata values already collected
in the linking document.