Optimize finding tags/relations with/without attributes (#470)
Attribute names and values are now better delineated in the indexed terms (there's a special character before the name, and before and after the value, so we can match exactly on either of these without any unwanted extra matches).

Tags/relations with attributes are now indexed twice: once without attributes and once with. If your query doesn't filter on any attributes, the former term is searched for; if it does include filters, we search for the latter. This speeds up searches without attribute filters if there are many different attribute values.

Changes how tags/relations are indexed but should gracefully deal with older and newer indexes because of an index metadata flag. (index metadata flags are a new addition and can be used to quickly add temporary properties to the index metadata to deal with this kind of variation without adding a whole new index format version)

We now strip any namespace prefixes from attributes while indexing. The BCQL parser doesn't allow colon in attribute name (and never has), so we either have to strip them or change the parser. We have chosen to strip them for now, with the slight risk of attribute name collisions (but keeping arbitrary prefixes that used to refer to namespaces doesn't necessarily make sense either).

There was also a bug where an invalid payload was added for an open tag, then later patched when we encounter te close tag, doing unnecessary work.

We no longer include span/relation attributes in their match info. The problem is that we index relations and tags twice now, once with and once without attributes, to improve search speed. If we always wanted to include attribute values in the match info, we'd have to store them in the payload, which would require a lot of extra space for too little (?) gain.