Fix BLSpanOrQuery issue by moving it to Lucene's package.
This gives it access to the package-private class used by SpanOrQuery.

Might seem like a hack, but these classes are so intertwined with how
Lucene works internally that they might as well live in the same package.

Arguably most subclasses of BLSpanQuery and BLSpans should be moved to
the same package eventually.

We might need to still double-check that BLSpanTermQuery and BLSpanMultiTermQueryWrapper
are up to date with their Lucene 9 counterparts.