blacklab-server/shared-with-me endpoint.
This will list private user corpora that have been shared
with you, the currently logged-in user. Only applies if
you're allowing users to create and share their own corpora.

To make this possible, a .userId file will now be saved in
each user corpora directory. If this file does not exist, it is
created once when the user logs in. So user corpora directories
created with older BlackLab versions not may have a .userId file
yet, and therefore cannot be reported in the /shared-with-me
response (because we can't tell the exact user id from the
directory name, because of name mangling).

Squashed commit of the following:

commit 98afc0f9cf62386005f46079645df0668a7fb448
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Thu Nov 23 10:37:42 2023 +0100

    Fix synchronization.

commit 4228cb35346ede5d6fd6430d7f03f652e5764b75
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Oct 31 15:50:18 2023 +0100

    Fix NPE when no userCollectionsDir.

commit cd71f1c48391f90f2c51e88fb7499172d6f7b635
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Oct 31 15:45:10 2023 +0100

    User id file. Load all user corpora.

commit ef8ba249af4593cfb5ead8c135df8bf85cc055a9
Author: Jan Niestadt <jan.niestadt@ivdnt.org>
Date:   Tue Oct 31 13:57:41 2023 +0100

    WIP shared-with-me endpoint.

    Does not yet take into account that private user corpora
    are not loaded until specifically requested, so it may miss
    corpora shared with you.

    We should find all corpora on startup and take note of what
    users they are shared with, without opening these corpora.