Fix indexFieldAs configuration option.
This option provides a simple way to index a field captured
with forEach under a different name instead. But the name
translation was applied too late in the process, which caused
issues with CGN indexing for the PublicationYear field.

Now we apply the name translation as soon as the metadata value
is added during indexing.