Allow authOIDC to return a uma-enabled user
Uma is not fully used yet, newly created corpora are not registered yet.
We will also need to implement a migration step that registers all current corpora and share lists with the UMA provider. There will also need to be made some decisions and template files to configure the keycloak server fully, such as formalizing the permission names and roles.