Work on token validation. Seems to work?
Have to override validation of pac4j however, because it expects an ID token and tries to verify the 'azp' parameter, which must include 'blacklab' (the clientId). Not ideal.
A further refinement would be to skip the entire profile creation in pac4j if the access token is a JWT, and only use the pac4j code if it's an opaque access token. Then the validation would become a non-issue, as info returned from the introspection endpoint is implicitly trusted and pac4j won't validate it. This would however require us to build our own profile mapping.